tests/wpt/tests/html/canvas/element/path-objects/2d.path.roundrect.nonfinite.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.roundrect.nonfinite</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.roundrect.nonfinite</h1>
<p class="desc">roundRect() with Infinity/NaN is ignored</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("roundRect() with Infinity/NaN is ignored");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50)
  ctx.moveTo(0, 0);
  ctx.lineTo(100, 0);
  ctx.roundRect(Infinity, 50, 1, 1, [0]);
  ctx.roundRect(-Infinity, 50, 1, 1, [0]);
  ctx.roundRect(NaN, 50, 1, 1, [0]);
  ctx.roundRect(0, Infinity, 1, 1, [0]);
  ctx.roundRect(0, -Infinity, 1, 1, [0]);
  ctx.roundRect(0, NaN, 1, 1, [0]);
  ctx.roundRect(0, 50, Infinity, 1, [0]);
  ctx.roundRect(0, 50, -Infinity, 1, [0]);
  ctx.roundRect(0, 50, NaN, 1, [0]);
  ctx.roundRect(0, 50, 1, Infinity, [0]);
  ctx.roundRect(0, 50, 1, -Infinity, [0]);
  ctx.roundRect(0, 50, 1, NaN, [0]);
  ctx.roundRect(0, 50, 1, 1, [Infinity]);
  ctx.roundRect(0, 50, 1, 1, [-Infinity]);
  ctx.roundRect(0, 50, 1, 1, [NaN]);
  ctx.roundRect(0, 50, 1, 1, [Infinity,0]);
  ctx.roundRect(0, 50, 1, 1, [-Infinity,0]);
  ctx.roundRect(0, 50, 1, 1, [NaN,0]);
  ctx.roundRect(0, 50, 1, 1, [0,Infinity]);
  ctx.roundRect(0, 50, 1, 1, [0,-Infinity]);
  ctx.roundRect(0, 50, 1, 1, [0,NaN]);
  ctx.roundRect(0, 50, 1, 1, [Infinity,0,0]);
  ctx.roundRect(0, 50, 1, 1, [-Infinity,0,0]);
  ctx.roundRect(0, 50, 1, 1, [NaN,0,0]);
  ctx.roundRect(0, 50, 1, 1, [0,Infinity,0]);
  ctx.roundRect(0, 50, 1, 1, [0,-Infinity,0]);
  ctx.roundRect(0, 50, 1, 1, [0,NaN,0]);
  ctx.roundRect(0, 50, 1, 1, [0,0,Infinity]);
  ctx.roundRect(0, 50, 1, 1, [0,0,-Infinity]);
  ctx.roundRect(0, 50, 1, 1, [0,0,NaN]);
  ctx.roundRect(0, 50, 1, 1, [Infinity,0,0,0]);
  ctx.roundRect(0, 50, 1, 1, [-Infinity,0,0,0]);
  ctx.roundRect(0, 50, 1, 1, [NaN,0,0,0]);
  ctx.roundRect(0, 50, 1, 1, [0,Infinity,0,0]);
  ctx.roundRect(0, 50, 1, 1, [0,-Infinity,0,0]);
  ctx.roundRect(0, 50, 1, 1, [0,NaN,0,0]);
  ctx.roundRect(0, 50, 1, 1, [0,0,Infinity,0]);
  ctx.roundRect(0, 50, 1, 1, [0,0,-Infinity,0]);
  ctx.roundRect(0, 50, 1, 1, [0,0,NaN,0]);
  ctx.roundRect(0, 50, 1, 1, [0,0,0,Infinity]);
  ctx.roundRect(0, 50, 1, 1, [0,0,0,-Infinity]);
  ctx.roundRect(0, 50, 1, 1, [0,0,0,NaN]);
  ctx.roundRect(Infinity, Infinity, 1, 1, [0]);
  ctx.roundRect(Infinity, Infinity, Infinity, 1, [0]);
  ctx.roundRect(Infinity, Infinity, Infinity, Infinity, [0]);
  ctx.roundRect(Infinity, Infinity, Infinity, Infinity, [Infinity]);
  ctx.roundRect(Infinity, Infinity, Infinity, 1, [Infinity]);
  ctx.roundRect(Infinity, Infinity, 1, Infinity, [0]);
  ctx.roundRect(Infinity, Infinity, 1, Infinity, [Infinity]);
  ctx.roundRect(Infinity, Infinity, 1, 1, [Infinity]);
  ctx.roundRect(Infinity, 50, Infinity, 1, [0]);
  ctx.roundRect(Infinity, 50, Infinity, Infinity, [0]);
  ctx.roundRect(Infinity, 50, Infinity, Infinity, [Infinity]);
  ctx.roundRect(Infinity, 50, Infinity, 1, [Infinity]);
  ctx.roundRect(Infinity, 50, 1, Infinity, [0]);
  ctx.roundRect(Infinity, 50, 1, Infinity, [Infinity]);
  ctx.roundRect(Infinity, 50, 1, 1, [Infinity]);
  ctx.roundRect(0, Infinity, Infinity, 1, [0]);
  ctx.roundRect(0, Infinity, Infinity, Infinity, [0]);
  ctx.roundRect(0, Infinity, Infinity, Infinity, [Infinity]);
  ctx.roundRect(0, Infinity, Infinity, 1, [Infinity]);
  ctx.roundRect(0, Infinity, 1, Infinity, [0]);
  ctx.roundRect(0, Infinity, 1, Infinity, [Infinity]);
  ctx.roundRect(0, Infinity, 1, 1, [Infinity]);
  ctx.roundRect(0, 50, Infinity, Infinity, [0]);
  ctx.roundRect(0, 50, Infinity, Infinity, [Infinity]);
  ctx.roundRect(0, 50, Infinity, 1, [Infinity]);
  ctx.roundRect(0, 50, 1, Infinity, [Infinity]);
  ctx.roundRect(0, 0, 100, 100, [new DOMPoint(10, Infinity)]);
  ctx.roundRect(0, 0, 100, 100, [new DOMPoint(10, -Infinity)]);
  ctx.roundRect(0, 0, 100, 100, [new DOMPoint(10, NaN)]);
  ctx.roundRect(0, 0, 100, 100, [new DOMPoint(Infinity, 10)]);
  ctx.roundRect(0, 0, 100, 100, [new DOMPoint(-Infinity, 10)]);
  ctx.roundRect(0, 0, 100, 100, [new DOMPoint(NaN, 10)]);
  ctx.roundRect(0, 0, 100, 100, [{x: 10, y: Infinity}]);
  ctx.roundRect(0, 0, 100, 100, [{x: 10, y: -Infinity}]);
  ctx.roundRect(0, 0, 100, 100, [{x: 10, y: NaN}]);
  ctx.roundRect(0, 0, 100, 100, [{x: Infinity, y: 10}]);
  ctx.roundRect(0, 0, 100, 100, [{x: -Infinity, y: 10}]);
  ctx.roundRect(0, 0, 100, 100, [{x: NaN, y: 10}]);
  ctx.lineTo(100, 50);
  ctx.lineTo(0, 50);
  ctx.fillStyle = '#0f0';
  ctx.fill();
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 90,45, 0,255,0,255);

});
</script>
