LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.roundrect.nonfinite-expected.txt

2d.path.roundrect.nonfinite
roundRect() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS roundRect() with Infinity/NaN is ignored
